Solution to Count Visited Nodes in a Directed Graph ( LeetCode 2876 ) Looked through LeetCode and came across the problem above. It doesn't have Editorial solution and it doesn't seem like a well-known problem from major lists that people use to prepare for interviews.
I wasn't able to find the problem that wouldn't have any solution online. I could try to solve some problems from online assesments that I got in the past, but that would violated NDA, so I opted to go with this one. Other option is to take the problems that just appeared in the some contests ( like the ones LeetCode hosts ), but people on YouTube post solutions to them within a day or two.

If you have any suggestions on where to find such problems, I would appreciate that. This time you will have to take my word that I didn't reference any solutions that are publicly available.