Solution to Paint House III ( LeetCode 1473 ) Came cross this problem while I was practicing DP problems. It's a known problem with solutions posted online, but the implementation is mine - like the last problem, I'm going to do it on my own.
Initially was going to propose Regular Expression Matching, but after scratching my head I was not able to solve it myself - only was able to come up with some ideas that were close to the solution, but not close enough to pass all test cases on LeetCode.

Anyway, I think this 'Paint' problem is quite interesting implementation wise, so that's why I chose it. Feel free to take a look at the problem description, I will ping you for review here when I'm done with solving it.